# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CAB230: Web Computing API",
    "description": "Express API based on the CAB230 Web Computing API by Liam Brock.",
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": { "email": "liam.brock99@gmail.com" },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://localhost:443"
    }
  ],
  "basePath": "/api",
  "tags": [
    {
      "name": "Authentication"
    },
    {
      "name": "Search"
    },
    {
      "name": "Helpers"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "description": "Register",
        "operationID": "post_register",
        "produces": ["application/json"],
        "requestBody": {
          "required": "true",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "yay! you've successfully registered your user account :)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "oops! it looks like that user already exists :("
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "description": "Login",
        "operationID": "post_login",
        "produces": ["application/json"],
        "requestBody": {
          "required": "true",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjo1Nzg2LCJlbWFpbCI6ImRlbW8xMjM0QGdtYWlsLmNvbSJ9LCJpYXQiOjE1NTUzMjQ2MDIsImV4cCI6MTU1NTQxMTAwMn0.mo85iKuAQ_h8n8-6pkeNNs5CCnV0G8hScxhi_ml_Q6c"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "Bearer"
                    },
                    "expires_in": {
                      "type": "integer",
                      "example": "86400"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "invalid login - bad password"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": ["Search"],
        "description": "Login",
        "operationID": "get_search",
        "produces": ["application/json"],
        "security": {
          "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "offence",
            "description": "The offence to search for",
            "schema": {
              "type": "string"
            },
            "required": "true"
          },
          {
            "in": "query",
            "name": "area",
            "description": "List of comma seperated areas to filter search results by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "age",
            "description": "List of comma seperated ages to filter search results by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "gender",
            "description": "List of comma seperated genders to filter search results by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "description": "List of comma seperated years to filter search results by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "description": "List of comma seperated genders to filter search results by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "object",
                      "properties": {
                        "offence": {
                          "type": "string",
                          "example": "Homicide (Murder)"
                        }
                      }
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "LGA": {
                            "type": "string",
                            "example": "Aurukun Shire Council"
                          },
                          "total": {
                            "type": "integer",
                            "example": "4"
                          },
                          "lat": {
                            "type": "number",
                            "example": "-27.470812"
                          },
                          "lng": {
                            "type": "number",
                            "example": "153.022455"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "oops! it looks like you're missing the offence query param"
                    }
                  },
                  "example": {
                    "message": "oops! it looks like you're missing the offence query param"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "oh no! it looks like your authorization token is invalid..."
                    }
                  },
                  "example": {
                    "message": "oh no! it looks like your authorization token is invalid..."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/offences": {
      "get": {
        "tags": ["Helpers"],
        "description": "Full listing of offences",
        "operationId": "get_offences",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "offences": {
                      "type": "array",
                      "example": [
                        "Assault",
                        "Armed Robbery",
                        "Advertising Prostitution"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/areas": {
      "get": {
        "tags": ["Helpers"],
        "description": "Full listing of areas",
        "operationId": "get_areas",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "areas": {
                      "type": "array",
                      "example": [
                        "Brisbane City Council",
                        "Gold Coast City Council"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ages": {
      "get": {
        "tags": ["Helpers"],
        "description": "Full listing of ages",
        "operationId": "get_ages",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ages": {
                      "type": "array",
                      "example": ["Adult", "Juvenile"],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/genders": {
      "get": {
        "tags": ["Helpers"],
        "description": "Full listing of genders",
        "operationId": "get_genders",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "genders": {
                      "type": "array",
                      "example": ["Male", "Female", "Not stated"],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/years": {
      "get": {
        "tags": ["Helpers"],
        "description": "Full listing of years",
        "operationId": "get_years",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "years": {
                      "type": "array",
                      "example": [2001, 2002, 2003, 2004],
                      "items": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/months": {
      "get": {
        "tags": ["Helpers"],
        "description": "Full listing of months",
        "operationId": "get_months",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "months": {
                      "type": "array",
                      "example": [1, 2, 3, 4],
                      "items": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}